requirements.txt
# Google ADK and AI dependencies
google-genai
google-generativeai

# Vector database and embeddings
chromadb
sentence-transformers

# File processing
python-magic
PyPDF2
python-docx
Pillow

# CLI interface
rich
click

# Data persistence
sqlalchemy

# Utilities
python-dotenv
pydantic

# Development
pytest
pytest-asyncio
black
